requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
cachetools==5.5.0
certifi==2023.7.22
cffi==1.17.1
charset-normalizer==3.2.0
coverage==7.6.7
cryptography==43.0.3
defusedxml==0.8.0rc2
dj-database-url==2.2.0
Django==5.1.3
django-cors-headers==4.6.0
django-environ==0.11.2
django-filter==24.3
django-markdown==0.8.4
django-oauth-toolkit==2.4.0
django-redis==5.4.0
django-rest-framework-social-oauth2==1.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
Faker==33.1.0
flake8==7.1.1
gunicorn==23.0.0
httplib2==0.22.0
idna==3.4
inflection==0.5.1
jwcrypto==1.5.6
Markdown==3.7
mccabe==0.7.0
oauthlib==3.2.2
packaging==23.1
pillow==11.0.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.8.0
pyotp==2.9.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.1
redis==5.2.0
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.0.4
whitenoise==6.7.0

"requests" 
